requirements.txt
pandas
bs4
Faker
requests
tqdm
pillow
python-dotenv
scraperwiki
dask[dataframes]